dded exec to run on all steps along with bug fixes and some refactoring.
This commit should be basically work for any experiment after the lcoations and
directory structure are copied. On the future to-do would be to move a portion,
possibly a significant portion the step_param files to global to make
editing and running easier. A lot of the information is replicated especially
since the instances are run from **relative** paths making the fileNames
effectively the same for most of the steps.

Additional ToDo:
- add aboslute timestamp of experiment step as attribute in analysis_abc
  and have it incorporated into global gel analysis as column in hdf file.
- add method to add timeStamp for each step time when particle center was scanned
  (simple remapping of z-height in imageStack)